Major rewrite
This is almost a complete rewrite.  It's not compatible with the old
version.

Instead of the -n option, there is now a -l option which specifies a
testing level.  It uses a logrithmic scale, where level 6 will take
about twice as long as level 5.

I added a -g option to turn off garbage collection during the
benchmarks.  By default garbage collection is on.  All benchmarks now
test memory usage in addition to performance.

You can now specify multiple configuration to test at once, and it
will spawn one process per configuration.

The benchmark output has changed.  The output is now
analysis-friendly CSV.  A results aggregator was also added that
allows easy comparison of differing configurations for the same
benchmark.

The README rewritten to give a better explanation for how to
configure and use this tool.

The lib and spec directories were removed, with all the benchmarking
logic being in the simple_orm_benchmarker.rb file and the models/*.rb
files.

Many additional benchmarks were added.  The individual benchmarks
are now a three element array of lambdas, with the following
elements:

Name/Label: Name of the benchmark
Before: Code to run before the benchmark
Benchmark: Code to benchmark