Add threading tests, -a flag for testing all configurations
I've added two types of threading tests, a lightly threaded test
and a heavily threaded test.  The both test the same code, but the
lightly threaded test uses a handful of threads and the heavily
threaded test uses a whole bunch (6 for the lightly threaded and
32 for the heavily threaded by default).

ActiveRecord doesn't clean up connections automatically, so there
is special support to automatically clean up connections when
a thread is finished.

I've also added a -a flag that tests all active configs, so you
don't need to specify all of them on the command line.

Thread testing doesn't work well in transactions, so I've added
a NO_TRANSACTION_BENCHES array for the threading tests.

While here, set the default database for sqlite3 to be file based
instead of memory based.  This makes for a more fair comparison
to other databases, and allows the threaded tests to work on
ActiveRecord.