Add MIT-LICENSE
While this benchmark was based on sequel-benchmark by Tim Lindquist,
it was mostly rewritten in 2009 and very little remains the same,
other than the use of Party and People as the models.

This didn't originally have a license, but I'm guessing that was
an oversight on Tim's part.  If Tim wants to use a different
license, I'll certainly be amenable to that.